requirements.txt
requests~=2.32.4
pandas~=2.3.1
python-dotenv~=1.1.1
#psycopg2~=2.9.10
psycopg2-binary~=2.9.10
flask~=3.1.2
gunicorn~=23.0.0